src/requirements.txt
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Deprecated==1.2.13
elasticsearch==7.16.3
fastapi==0.72.0
frozenlist==1.2.0
h11==0.12.0
idna==3.3
multidict==5.2.0
orjson==3.6.5
packaging==21.3
psycopg2-binary==2.9.3
pydantic==1.9.0
pyparsing==3.0.6
python-dotenv==0.19.2
redis==4.1.1
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0
wrapt==1.13.3
yarl==1.7.2
